requirements.txt
streamlit
librosa
numpy
tempfile2
openai>=1.0.0
python-dotenv
elevenlabs
